refactor: вынести сообщения об ошибках в отдельный файл
