+++ src/utils/constants/constants.js
@@ -1,0 +1,8 @@
+const constants = {
+  serverErrorMessage: `Во время запроса произошла ошибка.
+      Возможно, проблема с соединением или сервер недоступен.
+      Подождите немного и попробуйте ещё раз».`,
+  requirementKeyword: 'Нужно ввести ключевое слово',
+};
+
+export default constants;
